Renaming utility module to avoid cyclic dependencies
